Added and hero section and applicable elements
I removed the hero image element and changed it to a background image to make it responsive for all devices. I also added 3 new div elements. The hero-header-container was created to make it easier to position the hero text properly so it's always centered vertically and horizontally. The hero-cover is the red overlay section and the hero-bottom is the large red bottom border.