Updated styles for hero section and sidebar
I started off by converting the main.css file to SCSS on my machine to make is easier and more organized to style so the css file reflects that. I styled all elements for the hero section as well as creating some basic mixins to style and position main elements such as the content wrapper and the sidebar.